+++ app/src/main/java/woowacourse/paint/view/canvas/CanvasView.kt
@@ -1,10 +1,14 @@
-package woowacourse.paint.view
+package woowacourse.paint.view.canvas
 
 import android.content.Context
 import android.graphics.Canvas
 import android.util.AttributeSet
 import android.view.MotionEvent
 import android.view.View
+import woowacourse.paint.view.canvas.paintHistory.PainterHistory
+import woowacourse.paint.view.palette.color.PaletteColor
+import woowacourse.paint.view.palette.PaletteMode
+import woowacourse.paint.view.palette.shape.PaletteShape
 
 class CanvasView : View {
     constructor(context: Context) : super(context)
