+++ app/src/main/java/woowacourse/paint/view/canvas/painter/CirclePainter.kt
@@ -1,10 +1,13 @@
-package woowacourse.paint.view
+package woowacourse.paint.view.canvas.painter
 
 import android.graphics.Canvas
 import android.graphics.Paint
 import android.graphics.RectF
+import woowacourse.paint.common.softPainter
+import woowacourse.paint.view.palette.color.PaletteColor
+import woowacourse.paint.view.palette.shape.PaletteShape
 
-data class RectanglePainter(
+data class CirclePainter(
     private val shape: PaletteShape,
     private val paint: Paint = DEFAULT_PAINT,
 ) : Painter {
@@ -16,7 +19,7 @@ data class RectanglePainter(
         paint = updatePaint(paintColor = paletteColor.color),
     )
 
-    override fun setThickness(thickness: Float): RectanglePainter = copy(
+    override fun setThickness(thickness: Float): CirclePainter = copy(
         paint = updatePaint(thickness = thickness),
     )
 
@@ -31,7 +34,6 @@ data class RectanglePainter(
     override fun onActionDown(x: Float, y: Float) {
         startX = x
         startY = y
-        rect.set(startX, startY, x, y)
     }
 
     override fun onActionMove(x: Float, y: Float) {
@@ -39,7 +41,7 @@ data class RectanglePainter(
     }
 
     override fun draw(canvas: Canvas) {
-        canvas.drawRect(rect, paint)
+        canvas.drawArc(rect, 0F, 360F, true, paint)
     }
 
     override fun extract(): Painter = copy()
