+++ app/src/main/java/woowacourse/paint/view/canvas/painter/Painter.kt
@@ -1,6 +1,9 @@
-package woowacourse.paint.view
+package woowacourse.paint.view.canvas.painter
 
 import android.graphics.Canvas
+import woowacourse.paint.view.palette.color.PaletteColor
+import woowacourse.paint.view.palette.PaletteMode
+import woowacourse.paint.view.palette.shape.PaletteShape
 
 interface Painter {
     fun setPaletteColor(paletteColor: PaletteColor): Painter
