+++ app/src/main/java/woowacourse/paint/view/canvas/painter/RectanglePainter.kt
@@ -1,10 +1,13 @@
-package woowacourse.paint.view
+package woowacourse.paint.view.canvas.painter
 
 import android.graphics.Canvas
 import android.graphics.Paint
 import android.graphics.RectF
+import woowacourse.paint.common.softPainter
+import woowacourse.paint.view.palette.color.PaletteColor
+import woowacourse.paint.view.palette.shape.PaletteShape
 
-data class CirclePainter(
+data class RectanglePainter(
     private val shape: PaletteShape,
     private val paint: Paint = DEFAULT_PAINT,
 ) : Painter {
@@ -16,7 +19,7 @@ data class CirclePainter(
         paint = updatePaint(paintColor = paletteColor.color),
     )
 
-    override fun setThickness(thickness: Float): CirclePainter = copy(
+    override fun setThickness(thickness: Float): RectanglePainter = copy(
         paint = updatePaint(thickness = thickness),
     )
 
@@ -31,6 +34,7 @@ data class CirclePainter(
     override fun onActionDown(x: Float, y: Float) {
         startX = x
         startY = y
+        rect.set(startX, startY, x, y)
     }
 
     override fun onActionMove(x: Float, y: Float) {
@@ -38,7 +42,7 @@ data class CirclePainter(
     }
 
     override fun draw(canvas: Canvas) {
-        canvas.drawArc(rect, 0F, 360F, true, paint)
+        canvas.drawRect(rect, paint)
     }
 
     override fun extract(): Painter = copy()
