+++ app/src/main/java/woowacourse/paint/view/palette/shape/PaletteShape.kt
@@ -1,4 +1,4 @@
-package woowacourse.paint.view
+package woowacourse.paint.view.palette.shape
 
 import android.graphics.Canvas
 import android.graphics.Color
