+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <woowacourse.paint.view.CanvasView
+    <woowacourse.paint.view.canvas.CanvasView
         android:id="@+id/canvas_view"
         android:layout_width="0dp"
         android:layout_height="0dp"
@@ -13,7 +13,7 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
-    <woowacourse.paint.view.PaletteView
+    <woowacourse.paint.view.palette.PaletteView
         android:id="@+id/palette_view"
         android:layout_width="0dp"
         android:layout_height="120dp"
